--- agpypeline/transformer.py
@@ -1,24 +1,0 @@
-"""Template class for a Transformer Algorithm
-"""
-import environment
-
-
-class Transformer:
-    """Class for containing a Transformer Algorithm
-    """
-
-    def __init__(self):
-        """Initializes a class instance
-        """
-
-    def perform_process(self, transformer: environment.Transformer, check_md: dict, transformer_md: dict,
-                        full_md: dict) -> dict:
-        """Perform the processing of data
-        Arguments:
-            transformer: instance of transformer class
-            check_md: dictionary
-            transformer_md: dictionary
-            full_md: dictionary
-        """
-        # pylint: disable=unused-argument
-        return {}
